LPS-89154 Update actions
